+++ config.json
@@ -2,17 +2,13 @@
     "port": 2500,
     "num_tcp_pieces": 3,
     "num_tls_pieces": 2,
-    "TCP_frag": 4,
     "tcp_sleep": 0.01,
-    "TLS_frag": 4,
     "fake_packet": "GET / HTTP/1.1\r\nHost: www.baidu.com\r\n\r\n",
     "fake_ttl": "query",
     "fake_sleep": 0.1,
     "my_socket_timeout": 180,
     "FAKE_ttl_auto_timeout": 2,
     "doh_server": "https://cloudflare-dns.com/dns-query",
-    "DNS_log_every": 5,
-    "TTL_log_every": 1,
     "mode": "TLSfrag",
     "enalbe_ipv6": false,
     "IPredirect": {
